Jacob: Implement: Express integrator: verify Express 4/5 forward compatibility and document
Task: https://app.clickup.com/t/86cb8p664
