+++ integrators/express/package.json
@@ -27,7 +27,9 @@
     "type-check": "tsc --noEmit",
     "build": "tsc",
     "pre-build": "tsc",
-    "test": "node --import tsx --test src/**/*.test.ts test/**/*.test.ts"
+    "test": "pnpm run test:express5 && pnpm run test:express4",
+    "test:express5": "node --import tsx --test src/**/*.test.ts test/middleware.test.ts",
+    "test:express4": "node --import tsx --test test/middleware.express4.test.ts"
   },
   "keywords": [
     "qelos",
@@ -57,9 +59,9 @@
     "@qelos/sdk": "workspace:*"
   },
   "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/express-serve-static-core": "^4.19.6",
-    "express": "^4.21.2",
+    "@types/express": "^5.0.0",
+    "express": "^5.0.0",
+    "express4": "npm:express@^4.21.2",
     "tsx": "^4.21.0",
     "typescript": "^5.6.3"
   }
