+++ integrators/express/test/middleware.express4.test.ts
@@ -1,0 +1,4 @@
+import express4 from 'express4';
+import { registerMiddlewareTests } from './middleware.suite.js';
+
+registerMiddlewareTests(express4, 'express 4');
